refactor: useGetCardInfo 훅 분리
